Restructured estimators Added agent logging
