+++ learnrl/estimators/__init__.py
@@ -1,0 +1,7 @@
+# LearnRL a python library to learn and use reinforcement learning
+# Copyright (C) 2020 Mathïs FEDERICO <https://www.gnu.org/licenses/>
+
+from learnrl.estimators.estimator import Estimator
+from learnrl.estimators.numpy import TableEstimator
+from learnrl.estimators.keras import KerasEstimator
+
