+++ learnrl/estimators/estimator.py
@@ -196,75 +196,3 @@ class Estimator():
     def __call__(self, observations, actions=None):
         return self.predict(observations, actions)
 
-class TableEstimator(Estimator):
-
-    def build(self, **kwargs):
-        dtype = kwargs.get('dtype', None)
-        self.name = 'table'
-        self.observation_encoder, self.observation_decoder = self._get_table_encoder_decoder(self.observation_space)
-        self.action_encoder, self.action_decoder = self._get_table_encoder_decoder(self.action_space)
-        self.table = np.zeros((self.observation_size, self.action_size), dtype=dtype)
-    
-    def load(self, data):
-        self.table = data
-
-    def fit(self, observations, actions, Y):
-        delta = Y - self.table[observations, actions]
-        self.table[observations, actions] += self.learning_rate * delta
-    
-    def predict(self, observations, actions):
-        if actions is not None:
-            return self.table[observations, actions]
-        else:
-            return self.table[observations, :]
-        
-    def __str__(self):
-        return str(self.table)
-
-class KerasEstimator(Estimator):
-
-    def __init__(self, observation_space, action_space,
-                       learning_rate=1e-3, learning_rate_decay=0,
-                       batch_size=32, freezed_steps=0, **kwargs):
-        self.model = None
-        super().__init__(observation_space, action_space, learning_rate, learning_rate_decay, **kwargs)
-        self.action_encoder, self.action_decoder = self._get_table_encoder_decoder(self.action_space)
-        self.freezed_steps = freezed_steps
-        if self.freezed_steps > 0:
-            self.step_freezed_left = freezed_steps
-            self.model_freezed = deepcopy(self.model)
-        self.name = 'keras'
-        self.batch_size = batch_size
-
-    def build(self, **kwargs):
-        raise NotImplementedError
-
-    def preprocess(self, observations, actions=None):
-        raise NotImplementedError
-
-    def fit(self, observations, actions, Y):
-       
-        x_train = self.preprocess(observations, actions)
-        y_train = self.model.predict(x_train)
-        y_train[np.arange(len(actions)), actions] = Y
-        self.model.fit(x_train, y_train, epochs=1, batch_size=self.batch_size, verbose=self.verbose)
-
-        if self.freezed_steps > 0:
-            if self.step_freezed_left == 0:
-                self.model_freezed.set_weights(self.model.get_weights()) 
-                self.step_freezed_left = self.freezed_steps
-                if self.verbose > 0:
-                    print("Freezed model updated")
-            self.step_freezed_left -= 1
-
-    def predict(self, observations, actions=None):
-        x = self.preprocess(observations, actions)
-        if self.freezed_steps > 0:
-            Y = self.model_freezed.predict(x)
-        else:
-            Y = self.model.predict(x)
-        if actions is not None:
-            return Y[actions]
-        else:
-            return Y
-
